requirements.txt
asgiref==3.8.1
asttokens==3.0.0
colorama==0.4.6
crispy-bootstrap5==2024.10
decorator==5.1.1
distlib==0.3.9
Django==5.1.5
django-crispy-forms==2.3
djangorestframework==3.15.2
executing==2.2.0
filelock==3.17.0
ipython==8.31.0
jedi==0.19.2
matplotlib-inline==0.1.7
mysqlclient==2.2.7
parso==0.8.4
phonenumbers==8.13.53
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
pure_eval==0.2.3
Pygments==2.19.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2024.2
virtualenv==20.29.1
wcwidth==0.2.13
